requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pandas==2.2.3
psycopg2-binary==2.9.9
